added tools to catch errors

+++ DataBase/db_utils.py
@@ -1,8 +1,33 @@
 from models import Session
 from flask import jsonify
 from functools import wraps
+from app import app
+import sqlalchemy
 
 session = Session()
+
+
+class InvalidUsage(Exception):
+    status_code = 400
+
+    def __init__(self, message, status_code=None, payload=None):
+        Exception.__init__(self)
+        self.message = message
+        if status_code is not None:
+            self.status_code = status_code
+        self.payload = payload
+
+    def to_dict(self):
+        rv = dict(self.payload or ())
+        rv['message'] = self.message
+        return rv
+
+
+@app.errorhandler(InvalidUsage)
+def handle_invalid_usage(error):
+    response = jsonify(error.to_dict())
+    response.status_code = error.status_code
+    return response
 
 
 def db_lifecycle(func):
@@ -19,8 +44,12 @@ def db_lifecycle(func):
                 return jsonify({'message': e.args[0], 'type': 'KeyError'}), 400
             elif isinstance(e, TypeError):
                 return jsonify({'message': e.args[0], 'type': 'TypeError'}), 400
+            elif isinstance(e, sqlalchemy.exc.IntegrityError):
+                return jsonify({'message': "duplicate unique value", 'type': 'IntegrityError'}), 400
+            # elif isinstance(e, sqlalchemy.exc.IntegrityError):
+            #     return jsonify({'message': "duplicate unique value", 'type': 'IntegrityError'}), 400
             else:
-                return jsonify({'message': str(e), 'type': 'InternalServerError'}), 500
+                raise e
 
     return wrapper
 
@@ -56,12 +85,16 @@ def get_entries(model_class, model_schema):  # GET all entries
 @db_lifecycle
 def get_entry_by_id(model_class, model_schema, id):  # GET entry by id
     entry = session.query(model_class).filter_by(id=id).first()
+    if entry is None:
+        raise InvalidUsage("Object not found", status_code=404)
     return jsonify(model_schema().dump(entry))
 
 
 @db_lifecycle
 def get_entry_by_username(model_class, model_schema, username):  # GET _user_ by username
     entry = session.query(model_class).filter_by(username=username).first()
+    if entry is None:
+        raise InvalidUsage("Object not found", status_code=404)
     return jsonify(model_schema().dump(entry))
 
 
@@ -69,6 +102,8 @@ def get_entry_by_username(model_class, model_schema, username):  # GET _user_ by
 @session_lifecycle
 def update_entry_by_id(model_class, model_schema, id, **kwargs):  # PUT entity by id
     entry = session.query(model_class).filter_by(id=id).first()
+    if entry is None:
+        raise InvalidUsage("Object not found", status_code=404)
     for key, value in kwargs.items():
         setattr(entry, key, value)
     return jsonify(model_schema().dump(entry))
@@ -78,19 +113,25 @@ def update_entry_by_id(model_class, model_schema, id, **kwargs):  # PUT entity b
 @session_lifecycle
 def delete_entry_by_id(model_class, model_schema, id):  # DELETE entity by id
     entry = session.query(model_class).filter_by(id=id).first()
+    if entry is None:
+        raise InvalidUsage("Object not found", status_code=404)
     session.delete(entry)
     return jsonify(model_schema().dump(entry))
 
 
-def get_entry_by_ids(model_class, user_id, auditorium_id):  # GET access by user and auditorium ids
-    session = Session()
-    return session.query(model_class).filter_by(user_id=user_id, auditorium_id=auditorium_id).first()
+@db_lifecycle
+def get_entry_by_ids(model_class, model_schema, user_id, auditorium_id):  # GET access by user and auditorium ids
+    entry = session.query(model_class).filter_by(user_id=user_id, auditorium_id=auditorium_id).first()
+    if entry is None:
+        raise InvalidUsage("Object not found", status_code=404)
+    return jsonify(model_schema().dump(entry))
 
 
-def delete_entry_by_ids(model_class, user_id, auditorium_id, commit=True):  # DELETE access by user and auditorium ids
-    session = Session()
-    model = session.query(model_class).filter_by(user_id=user_id, auditorium_id=auditorium_id).first()
-    session.delete(model)
-    if commit:
-        session.commit()
-    return model
+@db_lifecycle
+@session_lifecycle
+def delete_entry_by_ids(model_class, model_schema, user_id, auditorium_id):  # DELETE access by user and auditorium ids
+    entry = session.query(model_class).filter_by(user_id=user_id, auditorium_id=auditorium_id).first()
+    if entry is None:
+        raise InvalidUsage("Object not found", status_code=404)
+    session.delete(entry)
+    return jsonify(model_schema().dump(entry))
